Moving yolov8 out of pytorch
