+++ setup.py
@@ -232,9 +232,9 @@ def _setup_entry_points() -> Dict:
 
     entry_points["console_scripts"].extend(
         [
-            "sparseml.ultralytics.train=sparseml.pytorch.yolov8.train:main",
-            "sparseml.ultralytics.val=sparseml.pytorch.yolov8.val:main",
-            "sparseml.ultralytics.export=sparseml.pytorch.yolov8.export:main",
+            "sparseml.ultralytics.train=sparseml.yolov8.train:main",
+            "sparseml.ultralytics.val=sparseml.yolov8.val:main",
+            "sparseml.ultralytics.export=sparseml.yolov8.export:main",
         ]
     )
